Move price component to be with the rest of the listing components
