+++ app/javascript/bundles/Lifestyle/components/listing/BasicInfo.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Price from '../Price'
+import Price from './Price'
 import SqFt from './SqFt'
 
 const BasicInfo = ({ address, price, sqft }) => {
